Added the ability to delete warhouses in the side bar. Added put requests in the warehouse controller

+++ Spring Boot/project-1/src/main/java/com/skillstorm/demo/models/warehouse.java
@@ -60,7 +60,7 @@ public class warehouse {
 		return warehouse_id;
 	}
 
-	public void setWarehouse_id(int warehouse_id) {
+	public void setWarehouse_id(long warehouse_id) {
 		this.warehouse_id = warehouse_id;
 	}
